migrated complex and complex math, along with a few other math helper headers, over to the headeronly namespace.
